reduce duplicate code between ida_search.c and ida_search_via_graph.c
