+++ rubikscubennnsolver/RubiksCube555.py
@@ -3414,7 +3414,7 @@ class RubiksCube555(RubiksCube):
             # into phase2 we should look. This can save some cycles by aborting phase2 once we know
             # the phase1_2 solution will not be shorter than our current min_phase1_2_len.
             if min_phase1_2_len is not None:
-                max_ida_threshold = min_phase1_2_len - phase1_len + 1
+                max_ida_threshold = min_phase1_2_len - phase1_len
             else:
                 max_ida_threshold = None
