+++ rubikscubennnsolver/ida_search_core.c
@@ -1339,3 +1339,31 @@ unsigned char steps_on_opposite_faces_in_order(move_type prev_move, move_type mo
             exit(1);
     }
 }
+
+unsigned char invalid_prune(unsigned char cost_to_here, move_type *moves_to_here, unsigned int threshold) {
+    // Lw2 R2 D2 F U' Lw2 D2 Lw2 R2 Fw2 B Uw2 D Fw2
+    move_type move_seq[14];
+
+    move_seq[0] = Lw2;
+    move_seq[1] = R2;
+    move_seq[2] = D2;
+    move_seq[3] = F;
+    move_seq[4] = U_PRIME;
+    move_seq[5] = Lw2;
+    move_seq[6] = D2;
+    move_seq[7] = Lw2;
+    move_seq[8] = R2;
+    move_seq[9] = Fw2;
+    move_seq[10] = B;
+    move_seq[11] = Uw2;
+    move_seq[12] = D;
+    move_seq[13] = Fw2;
+
+    if (threshold == 15) {
+        if (memcmp(moves_to_here, move_seq, sizeof(move_type) * cost_to_here) == 0) {
+            return 1;
+        }
+    }
+
+    return 0;
+}
